rename chat_locks to chat_scheduler
