+++ src/renderer/stores/chat/chat_scheduler.ts
